Fix event

+++ MicroRabbit/MicroRabbit.Banking.Domain/Commands/CreateTransferCommand.cs
@@ -6,7 +6,7 @@ namespace MicroRabbit.Banking.Domain.Commands
 {
    public class CreateTransferCommand: TransferCommand
     {
-        public CreateTransferCommand(int from, int to, decimal amount)
+        public CreateTransferCommand(int from, int to, int amount)
         {
             From = from;
             To = to;
