+++ MicroRabbit/MicroRabbit.Banking.Domain/Models/Account.cs
@@ -8,6 +8,6 @@ namespace MicroRabbit.Banking.Domain.Models
     {
         public int Id { get; set; }
         public string AccountType { get; set; }
-        public decimal AccountBalance { get; set; }
+        public int AccountBalance { get; set; }
     }
 }
